// app/types.ts
export interface FoodItem {
  foodId: number;
  concatName: string;
  measurementType: string | null;
  calories: number;
  carbs: number;
  protein: number;
  fats: number;
  fibers: number;
  avgUnitSize: number | null;
  amount: number;
}
export interface FoodDataItem {
  foodId: number;
  concatName: string;
  measurementType: string | null;
  calories: number;
  carbs: number;
  protein: number;
  fats: number;
  fibers: number;
  avgUnitSize: number | null;
}